ingestion/producer/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
kafka-python==2.3.0
requests==2.32.5
urllib3==2.5.0
gtfs-realtime-bindings==1.0.0
pydantic==2.12.5